+++ app/src/main/java/com/shichuang/mobileworkingticket/adapter/ReportProductAdapter.java
@@ -22,7 +22,7 @@ public class ReportProductAdapter extends BaseQuickAdapter<ReportProductList.Rep
         helper.setText(R.id.tv_product_drawing_no, "产品图号：" + item.getProductDrawingNo());
         HorizontalTextProgressBar mProgressBar = helper.getView(R.id.progress_bar);
         // 比例
-        int progress = (int) ((float) item.getFinishCount() * 100 / item.getAllCount());
+        int progress = (int) (item.getFinishCount() * 100 / item.getAllCount());
         mProgressBar.setMax(100);
         mProgressBar.setProgress(progress);
         mProgressBar.setTextProcess(progress);
