+++ app/src/main/java/com/shichuang/mobileworkingticket/entify/ReportProductList.java
@@ -34,9 +34,9 @@ public class ReportProductList {
         @SerializedName("product_drawing_no")
         private String productDrawingNo;
         @SerializedName("finish_count")
-        private int finishCount;
+        private double finishCount;
         @SerializedName("all_count")
-        private int allCount;
+        private double allCount;
 
         public String getWorkOrderNo() {
             return workOrderNo;
@@ -54,19 +54,19 @@ public class ReportProductList {
             this.productDrawingNo = productDrawingNo;
         }
 
-        public int getFinishCount() {
+        public double getFinishCount() {
             return finishCount;
         }
 
-        public void setFinishCount(int finishCount) {
+        public void setFinishCount(double finishCount) {
             this.finishCount = finishCount;
         }
 
-        public int getAllCount() {
+        public double getAllCount() {
             return allCount;
         }
 
-        public void setAllCount(int allCount) {
+        public void setAllCount(double allCount) {
             this.allCount = allCount;
         }
     }
